+++ services/web/server/src/simcore_service_webserver/wallets/_groups_models.py
@@ -1,0 +1,24 @@
+from datetime import datetime
+
+from models_library.groups import GroupID
+from pydantic import BaseModel, ConfigDict
+
+
+class WalletGroupGetDB(BaseModel):
+    gid: GroupID
+    read: bool
+    write: bool
+    delete: bool
+    created: datetime
+    modified: datetime
+
+
+class WalletGroupGet(BaseModel):
+    gid: GroupID
+    read: bool
+    write: bool
+    delete: bool
+    created: datetime
+    modified: datetime
+
+    model_config = ConfigDict(from_attributes=True)
