+++ services/web/server/src/simcore_service_webserver/wallets/models.py
@@ -1,7 +1,6 @@
 """Domain models for wallets."""
 
-from ._groups_api import WalletGroupGet
-from ._groups_db import WalletGroupGetDB
+from ._groups_models import WalletGroupGet, WalletGroupGetDB
 from ._schemas import WalletsPathParams
 
 __all__: tuple[str, ...] = (
